Minor Changes
I'm kidding, I'm kidding! Though, the changes I made for this one were indeed minor. All I changed for this commit were a couple parts of the abilities (like the soul mark, I think) as well as replacing the player character in the correct spot, and changing it's ability back to the soul mark instead of the ground slam.